Math library refactoring working as expected now.
There was confusion between the distinction on Lines and Edges. For the
purpose of the BSP tree intersection testing, a Line is treated as an
infinite line (ie. not bounded), of the usual form y = ax +b, where-ass
the Edge is actually a bounded line on the plane.

This means the intersection tests are not TRUE intersection tests
between lines, but rather test whether an infite line has both points of
the Edge to the same side. If both points are to the same side, the
respective value is returned. If the points are on different sides, the
intersecetion is given. If both points are on the infinite line, then
they are classified as co-incident.

This clears up a LOT of confusion.