Added the concept of subtimers to the Timer utility.
Subtimers only print out timing information when the "parent" timer is
stopped. This allows us to time operations that are executed many times
per "logical step" (which we time with a timer). The timing info will be
accumulated for the subtimer and will be logged when .end() is called
for the containing timer.