Removed all non-doom related code.
Im going ahead with optimising the rasterizer/renderer/btrees/quadtrees
with the aim of creating a doom top down game. The aim is on getting the
rendering times down.

If we ever decide to go ahead with the GSM game, we will benefit from
the development on this branch as well.