Creating a backing canvas solves our Patterning problem!
By having a backing canvas, we can solve the patterning problem we are
having with normal Canvas pattern drawing.

The reason we are having the "masking" effect, is because we are doing
the "scrolling" of the canvas ourselves. So basically, we have a world,
say from [wx1,wy1] to [wx2, wy2]. Then we have a viewport, say [vx1,
vy1] to [vx2,vy2]. We also have a canvas, going from [0,0] to
[1024,1024].

Combining these three objects, gives us a projection of the part of the
world defined by the "viewport" onto the canvas, which is (1024,1024).
But the canvas itself aligns its textures to it's OWN (0,0) coordinate
(not our world's), and thus it seems as though the textures are
"masked".

The solution I have here is to creating a backing canvas, which is large
enough to hold the whole level. Then we first draw the viewport
[vx1,vy1]x[vx2,vy2] onto the backing viewport, also at
[vx1,vy1]x[vx2,vy2]. This means the textures are now correct, because
the canvas (0,0) is the world (0,0). Now we simply "copy" the subcanvas
[vx1,vy1]x[vx2,vy2] onto the "front" canvas, and voila!

It seems to work well, and the framerates seem very good. It seems to
get a bit unstable though after a while. Need to look into it.

Also, the size of the backing canvas is hardcoded for now.